pyproject.toml
[project]
name = "review-app"
version = "0.2.0"
description = "AI Code Review App — GitHub App + Plane integration + semantic indexer"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Async task queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Validation / settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # AI
    "anthropic>=0.34.0",

    # HTTP client
    "httpx>=0.27.0",

    # GitHub App auth
    "PyJWT>=2.9.0",
    "cryptography>=43.0.0",

    # Config
    "pyyaml>=6.0.2",

    # BM25 hybrid search (Gap 1 — GitNexus pattern)
    "rank-bm25>=0.2.2",

    # Phase 2: Tree-sitter semantic indexer
    # tree-sitter-languages bundles 100+ grammars as pre-compiled wheels
    # (no need to compile each grammar individually)
    "tree-sitter>=0.23.0",
    "tree-sitter-languages>=1.10.2",

    # Phase 2: Vector store + embeddings (all-MiniLM-L6-v2 runs locally, free)
    "chromadb>=0.5.0",

    # multipart for file uploads (future dashboard use)
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
